backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]
sqlalchemy[asyncio]
psycopg2-binary
pydantic-settings
python-dotenv
aiosqlite
requests